Rearrange fields and refactor countTotalFiles function
Modified the ordering of the `DownloadProgress` struct fields for better readability. Refactored `countTotalFiles` function to accept an array of strings as argument instead of the entire `DownloadLocations` object.
This change simplifies the function's signature and makes it more flexible for counting files for individual locations, as well as a full list of locations. This also fixes an inefficiency where the total files were being recalculated unnecessarily within the download loop.